refactor: Refactored wildfire subfolder
